+++ src/client/command.rs
@@ -87,11 +87,11 @@ impl<'a> CommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn set_controller_led(
+    pub fn set_controller_led<C: Into<Color>>(
         &mut self,
         controller_id: impl ControllerIndex,
         led_id: usize,
-        color: Color,
+        color: C,
     ) -> OpenRgbResult<()> {
         let cmd = self.get_cmd_mut(controller_id)?;
         cmd.set_led(led_id, color)
@@ -102,10 +102,10 @@ impl<'a> CommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn set_controller_leds(
+    pub fn set_controller_leds<C: Into<Color>>(
         &mut self,
         controller_id: impl ControllerIndex,
-        colors: impl IntoIterator<Item = Color>,
+        colors: impl IntoIterator<Item = C>,
     ) -> OpenRgbResult<()> {
         let cmd = self.get_cmd_mut(controller_id)?;
         cmd.set_leds(colors)
@@ -116,11 +116,11 @@ impl<'a> CommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn set_controller_zone_leds(
+    pub fn set_controller_zone_leds<C: Into<Color>>(
         &mut self,
         controller_id: impl ControllerIndex,
         zone_id: usize,
-        colors: impl IntoIterator<Item = Color>,
+        colors: impl IntoIterator<Item = C>,
     ) -> OpenRgbResult<()> {
         let cmd = self.get_cmd_mut(controller_id)?;
         cmd.set_zone_leds(zone_id, colors)
@@ -131,12 +131,12 @@ impl<'a> CommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn set_controller_zone_led(
+    pub fn set_controller_zone_led<C: Into<Color>>(
         &mut self,
         controller_id: impl ControllerIndex,
         zone_id: usize,
         led_idx: usize,
-        color: Color,
+        color: C,
     ) -> OpenRgbResult<()> {
         let cmd = self.get_cmd_mut(controller_id)?;
         cmd.set_zone_led(zone_id, led_idx, color)
@@ -147,12 +147,12 @@ impl<'a> CommandGroup<'a> {
     /// # Errors
     ///
     /// Returns an error if the controller is not found in this group.
-    pub fn set_controller_segment_leds(
+    pub fn set_controller_segment_leds<C: Into<Color>>(
         &mut self,
         controller_id: impl ControllerIndex,
         zone_id: usize,
         segment_id: usize,
-        colors: impl IntoIterator<Item = Color>,
+        colors: impl IntoIterator<Item = C>,
     ) -> OpenRgbResult<()> {
         let cmd = self.get_cmd_mut(controller_id)?;
         cmd.set_segment_leds(zone_id, segment_id, colors)
@@ -188,49 +188,55 @@ impl<'a> Command<'a> {
     }
 
     /// Adds a command to update a single LED in this controller.
-    pub fn set_led(&mut self, led_id: usize, color: Color) -> OpenRgbResult<()> {
-        self.add_command(SetLedCommand::Single { led_id, color })
+    pub fn set_led<C: Into<Color>>(&mut self, led_id: usize, color: C) -> OpenRgbResult<()> {
+        self.add_command(SetLedCommand::Single {
+            led_id,
+            color: color.into(),
+        })
     }
 
     /// Adds a command to update multiple LEDs in this controller.
-    pub fn set_leds(&mut self, colors: impl IntoIterator<Item = Color>) -> OpenRgbResult<()> {
+    pub fn set_leds<C: Into<Color>>(
+        &mut self,
+        colors: impl IntoIterator<Item = C>,
+    ) -> OpenRgbResult<()> {
         self.add_command(SetLedCommand::Controller {
-            colors: colors.into_iter().collect(),
+            colors: colors.into_iter().map(Into::into).collect(),
         })
     }
 
     /// Adds a command to update a single LED in a zone in this controller.
-    pub fn set_zone_led(
+    pub fn set_zone_led<C: Into<Color>>(
         &mut self,
         zone_id: usize,
         led_idx: usize,
-        color: Color,
+        color: C,
     ) -> OpenRgbResult<()> {
         self.add_command(SetLedCommand::Single {
             led_id: self.controller.get_zone_led_offset(zone_id)? + led_idx,
-            color,
+            color: color.into(),
         })
     }
 
     /// Adds a command to update multiple LEDs in a zone in this controller.
-    pub fn set_zone_leds(
+    pub fn set_zone_leds<C: Into<Color>>(
         &mut self,
         zone_id: usize,
-        colors: impl IntoIterator<Item = Color>,
+        colors: impl IntoIterator<Item = C>,
     ) -> OpenRgbResult<()> {
         self.add_command(SetLedCommand::Zone {
             zone_id,
-            colors: colors.into_iter().collect(),
+            colors: colors.into_iter().map(Into::into).collect(),
         })
     }
 
     /// Adds a command to update a single led in a segment in a zone in this controller.
-    pub fn set_segment_led(
+    pub fn set_segment_led<C: Into<Color>>(
         &mut self,
         zone_id: usize,
         segment_id: usize,
         led_idx: usize,
-        color: Color,
+        color: C,
     ) -> OpenRgbResult<()> {
         let led_id = led_idx
             + self
@@ -238,20 +244,23 @@ impl<'a> Command<'a> {
                 .get_zone(zone_id)?
                 .get_segment(segment_id)?
                 .offset();
-        self.add_command(SetLedCommand::Single { led_id, color })
+        self.add_command(SetLedCommand::Single {
+            led_id,
+            color: color.into(),
+        })
     }
 
     /// Adds a command to update multiple LEDs in a segment in a zone in this controller.
-    pub fn set_segment_leds(
+    pub fn set_segment_leds<C: Into<Color>>(
         &mut self,
         zone_id: usize,
         segment_id: usize,
-        colors: impl IntoIterator<Item = Color>,
+        colors: impl IntoIterator<Item = C>,
     ) -> OpenRgbResult<()> {
         self.add_command(SetLedCommand::Segment {
             zone_id,
             segment_id,
-            colors: colors.into_iter().collect(),
+            colors: colors.into_iter().map(Into::into).collect(),
         })
     }